feat: add EpisodesListButton component and integrate it into StreamEdit and StreamShow
